remove NO GUI profiles and obsolete GUI component and its tests
